.history/requirements_20250522125303.txt

# 🧱 Django 핵심 프레임워크
Django==5.2
sqlparse==0.5.3
asgiref==3.8.1
tzdata==2025.2
setuptools==80.4.0

# 🔐 인증 / 보안 / 환경 변수
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dotenv
cryptography==44.0.3
pyOpenSSL==25.0.0
service-identity==24.2.0
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
cffi==1.17.1
pycparser==2.22

# 🌐 CORS 및 HTTP 통신
django-cors-headers==4.7.0
httpx==0.27.0

# 💬 실시간 통신 (Channels + Redis)
channels==4.2.2
channels_redis==4.2.1
daphne==4.1.2
redis==6.0.0
django-redis==5.4.0

# 🧵 Twisted 관련 (비동기 네트워크)
Twisted==24.11.0
autobahn==24.4.2
txaio==23.1.1
attrs==25.3.0
Automat==25.4.16
incremental==24.7.2
constantly==23.10.4
hyperlink==21.0.0
zope.interface==7.2

# 🧠 GPT API 연동
openai==0.28.0

# 🖼️ 이미지 처리 / 업로드
pillow

# ☁️ Firebase 연동
firebase-admin==6.4.0

# 🐘 PostgreSQL 연동
psycopg[binary]==3.1.18

# 📦 기타 유틸
typing_extensions==4.13.2
msgpack==1.1.0

